+++ CHANGELOG.md
@@ -22,8 +22,8 @@ Extend `dfx ledger transfer` and `dfx ledger balance` to support [ICRC-1 standar
 
 ### feat: Improve 'dfx canister logs' with several options
 
-Improve 'dfx canister logs' with several options
-- `--tail` and `-N` to show the last N lines of the logs
+Improve `dfx canister logs` with several options
+- `--tail` to show the last number of the logs
 - `--since` to show the logs newer than a relative duration
 - `--since-time` to show the logs newer than a specific timestamp